+++ playwright.config.ts
@@ -14,12 +14,14 @@ if (process.env['ADMIN_URL']) {
 } else {
   process.env['ADMIN_URL'] = process.env['APP_URL'] + 'admin/';
 }
-
+if (process.env['ADMIN_API_URL']) {
+  process.env['ADMIN_API_URL'] = (process.env['ADMIN_API_URL'] ?? '').replace(/\/+$/, '') + '/';
+}
 if (!process.env['WEBSERVER_COMMAND']) {
-  if (process.env['WEBSERVER_COMMAND'] === defaultAppUrl) {
-    process.env['WEBSERVER_COMMAND'] = 'docker compose up --pull=always --quiet-pull shopware';
+  if (process.env['APP_URL'] === defaultAppUrl) {
+    process.env['WEBSERVER_COMMAND'] = 'docker compose up --pull=always --quiet-pull';
   } else {
-    process.env['WEBSERVER_COMMAND'] = 'sleep 1h';
+    process.env['WEBSERVER_COMMAND'] = 'sleep 999d';
   }
 }
 
@@ -27,9 +29,9 @@ export default defineConfig({
   testDir: './tests',
   fullyParallel: true,
   forbidOnly: !!process.env.CI,
-  timeout: 60000,
+  timeout: 120000,
   expect: {
-    timeout: 10_000,
+    timeout: 30_000,
   },
   retries: 0,
   workers: process.env.CI ? 2 : 1,
@@ -44,7 +46,7 @@ export default defineConfig({
   },
   // We abuse this to wait for the external webserver
   webServer: {
-    command: process.env['WEBSERVER_COMMAND'] ?? 'sleep 1h',
+    command: process.env['WEBSERVER_COMMAND'] ?? 'sleep 999d',
     url: process.env['APP_URL'],
     reuseExistingServer: true,
     timeout: 180000,