+++ src/page-objects/StorefrontPages.ts
@@ -23,6 +23,9 @@ import { CheckoutOrderEdit } from './storefront/CheckoutOrderEdit';
 import { AccountAddressCreate } from './storefront/AccountAddresssCreate';
 import { PageNotFound } from './storefront/PageNotFound';
 import { ContactForm } from './storefront/ContactForm';
+import { Wishlist } from './storefront/Wishlist';
+import { Footer } from './storefront/Footer';
+import { Header } from './storefront/Header';
 
 export interface StorefrontPageTypes {
     StorefrontHome: Home;
@@ -47,6 +50,9 @@ export interface StorefrontPageTypes {
     StorefrontCheckoutOrderEdit: CheckoutOrderEdit;
     StorefrontPageNotFound: PageNotFound;
     StorefrontContactForm: ContactForm;
+    StorefrontWishlist: Wishlist;
+    StorefrontFooter: Footer;
+    StorefrontHeader: Header;
 }
 
 export const StorefrontPageObjects = {
@@ -72,6 +78,9 @@ export const StorefrontPageObjects = {
     CheckoutOrderEdit,
     PageNotFound,
     ContactForm,
+    Wishlist,
+    Footer,
+    Header,
 }
 
 export const test = base.extend<FixtureTypes>({
@@ -162,5 +171,17 @@ export const test = base.extend<FixtureTypes>({
   
     StorefrontContactForm: async ({ StorefrontPage, InstanceMeta }, use) => {
         await use(new ContactForm(StorefrontPage, InstanceMeta));
+    },
+  
+    StorefrontWishlist: async ({ StorefrontPage }, use) => {
+        await use(new Wishlist(StorefrontPage));
+    },
+
+    StorefrontFooter: async ({ StorefrontPage }, use) => {
+        await use(new Footer(StorefrontPage));
+    },
+
+    StorefrontHeader: async ({ StorefrontPage }, use) => {
+        await use(new Header(StorefrontPage));
     },
 });