+++ src/page-objects/administration/CustomerGroupDetail.ts
@@ -1,6 +1,7 @@
 import type { Page, Locator } from '@playwright/test';
 import type { PageObject } from '../../types/PageObject';
 import { CustomerGroupCreate } from './CustomerGroupCreate';
+import { HelperFixtureTypes } from '../../fixtures/HelperFixtures';
 
 export class CustomerGroupDetail extends CustomerGroupCreate implements PageObject {
     public readonly headline: Locator;
@@ -8,8 +9,8 @@ export class CustomerGroupDetail extends CustomerGroupCreate implements PageObje
     public readonly technicalUrl: Locator;
     public readonly saleschannelUrl: Locator;
 
-    constructor(public readonly page: Page) {
-        super(page);
+    constructor(public readonly page: Page, public readonly instanceMeta: HelperFixtureTypes['InstanceMeta']) {
+        super(page, instanceMeta);
         this.headline = page.locator('.smart-bar__header');
         this.selectedSalesChannel = page.locator('.sw-select-selection-list');
         this.technicalUrl = page.getByLabel('Technical URL');