+++ src/page-objects/administration/FlowBuilderListing.ts
@@ -2,7 +2,6 @@ import type { Page, Locator } from '@playwright/test';
 import type { PageObject } from '../../types/PageObject';
 
 export class FlowBuilderListing implements PageObject {
-
     public readonly createFlowButton: Locator;
     public readonly firstFlowName: Locator;
     public readonly firstFlowContextButton: Locator;
@@ -16,6 +15,7 @@ export class FlowBuilderListing implements PageObject {
     public readonly flowDeleteButton: Locator;
     public readonly successAlert: Locator;
     public readonly successAlertMessage: Locator;
+    public readonly searchBar: Locator;
 
     constructor(public readonly page: Page) {
         this.createFlowButton = page.locator('.sw-flow-list__create');
@@ -33,6 +33,7 @@ export class FlowBuilderListing implements PageObject {
         this.downloadFlowButton = page.getByRole('button', { name: 'Download flow' });
         this.successAlert = page.locator('.sw-alert__body');
         this.successAlertMessage = page.locator('.sw-alert__message');
+        this.searchBar = page.locator('.sw-search-bar__input');
     }
 
     url() {
@@ -55,6 +56,6 @@ export class FlowBuilderListing implements PageObject {
             flowNameText: flowNameText,
             flowEventNameText: flowEventNameText,
             flowContextMenuButton: flowContextMenuButton,
-        }
+        };
     }
 }