+++ src/page-objects/administration/ProductDetail.ts
@@ -1,5 +1,8 @@
 import type { Page, Locator } from '@playwright/test';
 import type { PageObject } from '../../types/PageObject';
+import { satisfies } from 'compare-versions';
+import { HelperFixtureTypes } from '../../fixtures/HelperFixtures';
+import { getCustomFieldCardLocators } from './modules/CustomFieldCard';
 
 export class ProductDetail implements PageObject {
 
@@ -79,7 +82,7 @@ export class ProductDetail implements PageObject {
      */
     public readonly customFieldCard: Locator;
 
-    constructor(public readonly page: Page) {
+    constructor(public readonly page: Page, public readonly instanceMeta: HelperFixtureTypes['InstanceMeta']) {
 
         this.savePhysicalProductButton = page.getByRole('button', { name: 'Save' });
         this.saveButtonCheckMark = page.locator('.icon--regular-checkmark-xs');
@@ -99,9 +102,13 @@ export class ProductDetail implements PageObject {
         this.activeForAllSalesChannelsToggle = page.locator('.sw-field--product-active').getByRole('checkbox');
         this.tagsInput = page.locator('.sw-product-category-form__tag-field');
         this.saleChannelsInput = page.locator('.sw-product-detail__select-visibility');
-        
+
         // Labelling
-        this.releaseDateInput = page.locator('.sw-product-detail-base__labelling-card').locator('.form-control');
+        if (satisfies(instanceMeta.version, '<6.7')) {
+            this.releaseDateInput = page.locator('.sw-product-detail-base__labelling-card').locator('.form-control');
+        } else {
+            this.releaseDateInput = page.locator('.dp__input_readonly');
+        }
 
         // Media upload interactions
         this.uploadMediaButton = page.getByRole('button', { name: 'Upload file' });
@@ -120,20 +127,29 @@ export class ProductDetail implements PageObject {
         this.propertyGroupColor = this.variantsModal.getByText('Color').first();
         this.propertyGroupSize = this.variantsModal.getByText('Size').first();
         this.propertyOptionGrid = this.variantsModal.locator('.sw-property-search__tree-selection__option_grid');
-        this.propertyOptionColorBlue = this.propertyOptionGrid.getByLabel('Blue');
-        this.propertyOptionColorRed = this.propertyOptionGrid.getByLabel('Red');
-        this.propertyOptionColorGreen = this.propertyOptionGrid.getByLabel('Green');
-        this.propertyOptionSizeSmall = this.propertyOptionGrid.getByLabel('Small');
-        this.propertyOptionSizeMedium = this.propertyOptionGrid.getByLabel('Medium');
-        this.propertyOptionSizeLarge = this.propertyOptionGrid.getByLabel('Large');
+        this.propertyOptionColorBlue = this.propertyOptionGrid.getByRole('row', { name: 'Blue' }).getByRole('checkbox');
+        this.propertyOptionColorRed = this.propertyOptionGrid.getByRole('row', { name: 'Red' }).getByRole('checkbox');
+        this.propertyOptionColorGreen = this.propertyOptionGrid.getByRole('row', { name: 'Green' }).getByRole('checkbox');
+        this.propertyOptionSizeSmall = this.propertyOptionGrid.getByRole('row', { name: 'Small' }).getByRole('checkbox');
+        this.propertyOptionSizeMedium = this.propertyOptionGrid.getByRole('row', { name: 'Medium' }).getByRole('checkbox');
+        this.propertyOptionSizeLarge = this.propertyOptionGrid.getByRole('row', { name: 'Large' }).getByRole('checkbox');
 
         this.specificationsTabLink = page.getByRole('tab', { name: 'Specifications' });
-        this.customFieldCard = page.locator('.sw-card').getByText('Custom fields');
+        if (satisfies(instanceMeta.version, '<6.7')) {
+            this.customFieldCard = page.locator('.sw-card').getByText('Custom fields');
+        } else {
+            this.customFieldCard = page.locator('.mt-card').getByText('Custom fields');
+        }
     }
 
     async getCustomFieldSetCardContentByName(customFieldSetName: string): Promise<Record<string, Locator>> {
+        let customFieldCard: Locator;
+        if (satisfies(this.instanceMeta.version, '<6.7')) {
+            customFieldCard = this.page.locator('.sw-card').filter({ hasText: 'Custom fields' });
+        } else {
+            customFieldCard = this.page.locator('.mt-card').filter({ hasText: 'Custom fields' });
+        }
 
-        const customFieldCard = this.page.locator('.sw-card').filter({ hasText: 'Custom fields' });
         const customFieldSetTab = customFieldCard.getByText(customFieldSetName);
         const customFieldSetTabCustomContent = customFieldCard.locator(`.sw-custom-field-set-renderer-tab-content__${customFieldSetName}`);
 
@@ -145,5 +161,9 @@ export class ProductDetail implements PageObject {
 
     url(productId: string) {
         return `#/sw/product/detail/${productId}/base`
+    }
+
+    async getCustomFieldCardLocators(customFieldSetName: string, customFieldTextName: string) {
+        return getCustomFieldCardLocators(this.page, customFieldSetName, customFieldTextName, this.instanceMeta);
     }
 }