+++ src/page-objects/administration/RuleCreate.ts
@@ -1,5 +1,7 @@
 import type { Page, Locator } from '@playwright/test';
 import type { PageObject } from '../../types/PageObject';
+import { getSelectFieldListitem } from './modules/SelectFieldListitem';
+import { HelperFixtureTypes } from '../../fixtures/HelperFixtures';
 
 export class RuleCreate implements PageObject {
 
@@ -14,8 +16,9 @@ export class RuleCreate implements PageObject {
     public readonly saveButton: Locator;
     public readonly cancelButton: Locator;
     public readonly smartBarHeader: Locator;
+    public readonly valueNotAvailableTooltip: Locator;
 
-    constructor(public readonly page: Page) {
+    constructor(public readonly page: Page, public readonly instanceMeta: HelperFixtureTypes['InstanceMeta']) {
         this.nameInput = page.getByLabel('Name');
         this.priorityInput = page.getByLabel('Priority');
         this.descriptionInput = page.getByLabel('Description');
@@ -27,9 +30,13 @@ export class RuleCreate implements PageObject {
         this.saveButton = page.getByRole('button', { name: 'Save' });
         this.cancelButton = page.getByRole('button', { name: 'Cancel' });
         this.smartBarHeader = page.locator('.smart-bar__header');
+        this.valueNotAvailableTooltip = page.locator('.sw-tooltip');
     }
 
     url() {
         return `#/sw/settings/rule/create/base`
+    }
+    async getSelectFieldListitem(selectField: Locator, listItem: string) {
+        return getSelectFieldListitem(this.page, selectField, listItem, this.instanceMeta);
     }
 }