+++ src/services/Cache.ts
@@ -1,0 +1,6 @@
+import { type AdminApiContext } from './AdminApiContext';
+
+export const clearDelayedCache = async (adminApiContext: AdminApiContext): Promise<void> => {
+    await adminApiContext.delete('./_action/cache-delayed');
+};
+