+++ src/services/ShopwareDataHelpers.ts
@@ -1,7 +1,7 @@
 import { APIResponse } from '@playwright/test';
 import { AdminApiContext } from './AdminApiContext';
 import type { components } from '@shopware/api-client/admin-api-types';
-import { Promotion } from '../types/ShopwareTypes';
+import { Flow, FlowTemplate, Promotion } from '../types/ShopwareTypes';
 
 type Language = components['schemas']['Language'] & {
     id: string,
@@ -108,6 +108,8 @@ export const getPaymentMethodId = async (adminApiContext: AdminApiContext, handl
 /**
  * Gives the default shipping method back called Standard
  * @param adminApiContext - An AdminApiContext entity
+ *
+ * @deprecated - Use getShippingMethodId instead
  */
 export const getDefaultShippingMethodId = async (adminApiContext: AdminApiContext): Promise<string> => {
     const resp = await adminApiContext.post('search/shipping-method', {
@@ -117,6 +119,23 @@ export const getDefaultShippingMethodId = async (adminApiContext: AdminApiContex
                 type: 'equals',
                 field: 'name',
                 value: 'Standard',
+            }],
+        },
+    });
+
+    const result = (await resp.json()) as { data: { id: string; active: boolean }[]; total: number };
+
+    return result.data[0].id;
+};
+
+export const getShippingMethodId = async (name: string, adminApiContext: AdminApiContext): Promise<string> => {
+    const resp = await adminApiContext.post('search/shipping-method', {
+        data: {
+            limit: 1,
+            filter: [{
+                type: 'equals',
+                field: 'name',
+                value: name,
             }],
         },
     });
@@ -250,6 +269,58 @@ export const getMediaId = async (fileName: string, adminApiContext: AdminApiCont
     const result = (await resp.json()) as { data: { id: string }[]; total: number };
     return result.data[0].id;
 };
+
+export const getFlowTemplate = async (flowTemplateId: string, adminApiContext: AdminApiContext): Promise<FlowTemplate> => {
+    const flowTemplateResponse = await adminApiContext.post(`search/flow-template`, {
+        data: {
+            limit: 1,
+            filter: [{
+                type: 'equals',
+                field: 'id',
+                value: flowTemplateId,
+            }],
+        },
+    });
+    const result = (await flowTemplateResponse.json());
+    return result.data[0];
+};
+
+export const getFlow = async (flowId: string, adminApiContext: AdminApiContext): Promise<Flow> => {
+    const flowResponse = await adminApiContext.post(`search/flow`, {
+        data: {
+            limit: 1,
+            filter: [{
+                type: 'equals',
+                field: 'id',
+                value: flowId,
+            }],
+            associations: { sequences: {}  },
+        },
+    });
+    const result = (await flowResponse.json());
+    return result.data[0];
+};
+
+export const compareFlowTemplateWithFlow = async (flowId:string, flowTemplateId: string, adminApiContext: AdminApiContext): Promise<boolean> => {
+    const flowTemplateData: FlowTemplate = await getFlowTemplate(flowTemplateId, adminApiContext);
+    const flowData: Flow = await getFlow(flowId, adminApiContext);
+    // compare triggers
+    if (flowTemplateData.config.eventName != flowData.eventName){
+        return false;
+    }
+    // compare sequences
+    let i = 0;
+    for (const sequenceTemplate of flowTemplateData.config.sequences) {
+        if (sequenceTemplate.actionName != flowData.sequences[i].actionName){
+            return false;
+        }
+        if (JSON.stringify(sequenceTemplate.config) != JSON.stringify(flowData.sequences[i].config)){
+            return false;
+        }
+        i++;
+    }
+    return true;
+}
 
 export function extractIdFromUrl(url: string): string | null {
     const segments = url.split('/');